requirements.txt
fastapi
uvicorn
pydantic
cloudinary
sqlalchemy
python-multipart
django
Mako
asyncpg
alembic
fastapi-utils
typing-inspect
pyjwt
passlib
bcrypt